more refactoring int->int64

+++ tests/test_wba/test_wb_assembly/test_examples.py
@@ -9,7 +9,7 @@ from tests.test_wba.conftest import window
 
 def naive_stride_list(start, stop, duration, foot=None, **paras):
     """A window list full of identical strides."""
-    x = np.arange(start, stop + duration, duration)
+    x = np.arange(start, stop + duration, duration, dtype="int64")
     start_end = zip(x[:-1], x[1:])
 
     return pd.DataFrame.from_records(
